Apply PR suggestion: rename compose container
